Custom lexer FTW!
Now with much nicer error messages, including line+col number, and tree
position, and a bunch of old html5lib ParseErrors are now much more
specific.

Still lots of stuff TODO, and a few edge cases not decided yet,
etc.

The html5lib stuff is pulled out completely, and but compatibility with
that is still available, and tests to make sure it still works is
there... in theory then it should be possible to get some kind of XHTML
validation with this too one day - but it's not something I need.